worker/requirements.txt
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic-settings==2.2.1
playwright==1.44.0
pyyaml==6.0.1
openpyxl==3.1.2
jinja2==3.1.4
markdown==3.6
httpx==0.27.0
anthropic>=0.40.0
